+++ lib/paper_trail/attribute_serializers/object_attribute.rb
@@ -23,12 +23,17 @@ module PaperTrail
       # Modifies `attributes` in place.
       # TODO: Return a new hash instead.
       def alter(attributes, serialization_method)
-        # Don't serialize before values before inserting into columns of type
-        # `JSON` on `PostgreSQL` databases.
-        return attributes if object_col_is_json?
+        # for columns of type `JSON` on `PostgreSQL` databases.
+        # only serialize encrypted attributes
+        if object_col_is_json?
+          encrypted_attributes = @model_class.try(:encrypted_attributes)
+          return attributes if encrypted_attributes.blank?
+        end
 
         serializer = CastAttributeSerializer.new(@model_class)
         attributes.each do |key, value|
+          next if encrypted_attributes&.exclude? key.to_sym
+
           attributes[key] = serializer.send(serialization_method, key, value)
         end
       end