+++ CHANGELOG.md
@@ -11,6 +11,11 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Added
 
+- [#1420](https://github.com/paper-trail-gem/paper_trail/pull/1420) - Adds support
+  for Active Record Encryption in Postgres json and jsonb columns.
+  The value stored in the object and object_changes columns will now be encrypted
+  for the model's encrypted attributes.
+
 - [#1416](https://github.com/paper-trail-gem/paper_trail/pull/1416) - Adds a
   model-configurable option `synchronize_version_creation_timestamp` which, if
   set to false, opts out of synchronizing timestamps between `Version.created_at`